Configure Windows release tasks to include firmware in generated package
The tool performs three distinct operations:

- Check the board's operating system version, and if outdated flash the new version.
- Check the board's BLE firmware version, and if outdated flash the new version.
- Flash the user application binary.

For this reason, the release package must contain the operating system and BLE firmware files in addition to the tool
executable. These files are stored in subfolders under the `firmwares` folder of the repository.

Previously, the `zip` command used to generate the release packages for Windows hosts was not correctly configured to
recurse into the folder when generating the archive. This resulted in the release package only containing an empty
folder where the firmware files should have been located. This caused uploads to fail when the target board had an
outdated operating system or BLE firmware version.