Use more meaningful variable names in release workflow
A GitHub Actions workflow is used to automatically generate production builds of the project.

A separate build is generated for each of the target host types. This is done using a job matrix, which creates a
parallel run of the workflow job for each target. The matrix defines variables that provide the data that is specific to
each job.

The variable names used previously did not clearly communicate their nature:

Previously, the matrix variable for the task name was named "os". This did not clearly communicate the nature of the
variable, making it difficult for anyone not intimately familiar with the workings of the workflow to understand its
code.