Update workflow artifact actions in release workflow
The "Release" GitHub Actions workflow generates binaries for a range of target hosts. This is done by using a job matrix
that produces each build in a parallel job. GitHub Actions workflow artifacts are used to transfer the generated files
between sequential jobs in the workflow. The "actions/upload-artifact" and "actions/download-artifact" actions are used
for this purpose.

GitHub made a significant change in the format of the workflow artifacts and has dropped support for the old format used
by versions of the actions prior to 4.0.0. This caused the release workflow to fail:

```
Error: This request has been automatically failed because it uses a deprecated version of `actions/upload-artifact: v3`.
Learn more: https://github.blog/changelog/2024-04-16-deprecation-notice-v3-of-the-artifact-actions/
```

Due to the change in the artifacts system, some refactoring of the "Release" workflow was required to accommodate the
action bumps:

Previously, a single artifact was used for this purpose, with each of the parallel jobs uploading its own generated
files to that artifact. However, support for uploading multiple times to a single artifact was dropped in version 4.0.0
of the "actions/upload-artifact" action. So it is now necessary to use a dedicated artifact for each of the builds.
These can be downloaded in aggregate by using the artifact name globbing and merging features which were introduced in
version 4.1.0 of the "actions/download-artifact" action.